feat: create HomePageClient component for enhanced project showcase
- Introduced HomePageClient component to manage project showcase with a responsive carousel.
- Implemented fallback showcase items for better user experience when no projects are available.
- Integrated async data fetching for projects in the main HomePage component.
- Cleaned up and organized code for improved maintainability and readability.
